requirements.txt
Pillow==6.2.0
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16
